+++ lib/dartantic_ai.dart
@@ -1,1 +1,2 @@
 export 'src/dartantic_ai.dart';
+export 'src/json_schema_extension.dart';